AJSI-644: app takes a long time to reconnect
Part of the fix is to try to shut down AJ and the controller
client gracefully when the app is exiting. There may still be
scenarios where we cannot shutdown gracefully (e.g., if
the app is force stopped or crashes), and in these cases
the amount of time to reconnect would depend on the
controller client.

Signed-off-by: David Diplock <ddiplock@qce.qualcomm.com>
Change-Id: I9e4135ff40b91449c3482ac255b1d3f142495aee